Move title menu to separate file
