+++ webapp/src/components/backstage/playbook_runs/playbook_run/header.tsx
@@ -3,32 +3,22 @@
 
 import styled from 'styled-components';
 
-import React, {useState} from 'react';
-import {FormattedMessage, useIntl} from 'react-intl';
-import {useDispatch, useSelector} from 'react-redux';
-import {GlobalState} from 'mattermost-redux/types/store';
+import React from 'react';
+import {useIntl} from 'react-intl';
+import {useDispatch} from 'react-redux';
 
 import CopyLink from 'src/components/widgets/copy_link';
 import {showRunActionsModal} from 'src/actions';
-import {exportChannelUrl, finishRun, getSiteUrl} from 'src/client';
-import {TitleButton} from '../../playbook_editor/controls';
+import {getSiteUrl} from 'src/client';
 import {PlaybookRun, Metadata as PlaybookRunMetadata} from 'src/types/playbook_run';
-import DotMenu, {DropdownMenuItem} from 'src/components/dot_menu';
-import {SemiBoldHeading} from 'src/styles/headings';
 import Tooltip from 'src/components/widgets/tooltip';
 import {HeaderIcon} from '../playbook_run_backstage/playbook_run_backstage';
 
 import {ExpandRight} from 'src/components/backstage/playbook_runs/shared';
 import RunActionsModal from 'src/components/run_actions_modal';
 import {navigateToUrl} from 'src/browser_routing';
-import {copyToClipboard} from 'src/utils';
-import {useToasts} from '../../toast_banner';
-import {useAllowChannelExport} from 'src/hooks';
-import UpgradeModal from '../../upgrade_modal';
-import {AdminNotificationType} from 'src/constants';
-import {outstandingTasks} from 'src/components/modals/update_run_status_modal';
-import {modals} from 'src/webapp_globals';
-import {makeUncontrolledConfirmModalDefinition} from 'src/components/widgets/confirmation_modal';
+
+import {ContextMenu} from './context_menu';
 
 interface HeaderProps {
     playbookRun: PlaybookRun;
@@ -39,12 +29,10 @@ export const HeaderContainer = ({playbookRun, playbookRunMetadata}: HeaderProps)
     const dispatch = useDispatch();
     const {formatMessage} = useIntl();
 
-    // const {add: addToast} = useToasts();
-
     return (
         <Container>
             <Icon className={'icon-star'}/>
-            <TitleMenuImpl playbookRun={playbookRun}/>
+            <ContextMenu playbookRun={playbookRun}/>
             <HeaderButton
                 tooltipId={'run-actions-button-tooltip'}
                 tooltipMessage={formatMessage({defaultMessage: 'Run Actions'})}
@@ -86,104 +74,6 @@ export const HeaderContainer = ({playbookRun, playbookRunMetadata}: HeaderProps)
     );
 };
 
-interface TitleMenuProps {
-    playbookRun: PlaybookRun;
-}
-
-const TitleMenuImpl = ({playbookRun}: TitleMenuProps) => {
-    const dispatch = useDispatch();
-    const {formatMessage} = useIntl();
-    const {add: addToast} = useToasts();
-
-    //@ts-ignore plugins state is a thing
-    const exportAvailable = useSelector<GlobalState, boolean>((state) => Boolean(state.plugins?.plugins?.['com.mattermost.plugin-channel-export']));
-    const allowChannelExport = useAllowChannelExport();
-    const [showModal, setShowModal] = useState(false);
-
-    const onExportClick = () => {
-        if (!allowChannelExport) {
-            setShowModal(true);
-            return;
-        }
-
-        window.location.href = exportChannelUrl(playbookRun.channel_id);
-    };
-
-    return (
-        <>
-            <DotMenu
-                dotMenuButton={TitleButton}
-                placement='bottom-end'
-                icon={
-                    <>
-                        <Title>{playbookRun.name}</Title>
-                        <i className={'icon icon-chevron-down'}/>
-                    </>
-                }
-            >
-                <DropdownMenuItem
-                    onClick={() => {
-                        copyToClipboard(getSiteUrl() + '/playbooks/runs/' + playbookRun?.id);
-                        addToast(formatMessage({defaultMessage: 'Copied!'}));
-                    }}
-                >
-                    <FormattedMessage defaultMessage='Copy link'/>
-                </DropdownMenuItem>
-                <DropdownMenuItem
-                    onClick={() => dispatch(showRunActionsModal())}
-                >
-                    <FormattedMessage defaultMessage='Run actions'/>
-                </DropdownMenuItem>
-                <DropdownMenuItem
-                    disabled={!exportAvailable}
-                    disabledAltText={formatMessage({defaultMessage: 'Install and enable the Channel Export plugin to support exporting the channel'})}
-                    onClick={async () => {
-                        onExportClick();
-                    }}
-                >
-                    <FormattedMessage defaultMessage='Export channel log'/>
-                </DropdownMenuItem>
-                <DropdownMenuItem
-                    onClick={() => {
-                        const outstanding = outstandingTasks(playbookRun.checklists);
-                        let confirmationMessage = formatMessage({defaultMessage: 'Are you sure you want to finish the run?'});
-                        if (outstanding > 0) {
-                            confirmationMessage = formatMessage(
-                                {defaultMessage: 'There {outstanding, plural, =1 {is # outstanding task} other {are # outstanding tasks}}. Are you sure you want to finish the run?'},
-                                {outstanding});
-                        }
-
-                        const onConfirm = () => {
-                            finishRun(playbookRun.id);
-                        };
-
-                        dispatch(modals.openModal(makeUncontrolledConfirmModalDefinition({
-                            show: true,
-                            title: formatMessage({defaultMessage: 'Confirm finish run'}),
-                            message: confirmationMessage,
-                            confirmButtonText: formatMessage({defaultMessage: 'Finish run'}),
-                            onConfirm,
-                            // eslint-disable-next-line @typescript-eslint/no-empty-function
-                            onCancel: () => {},
-                        })));
-                    }}
-                >
-                    <FormattedMessage defaultMessage='Finish run'/>
-                </DropdownMenuItem>
-            </DotMenu>
-            <UpgradeModal
-                messageType={AdminNotificationType.EXPORT_CHANNEL}
-                show={showModal}
-                onHide={() => setShowModal(false)}
-            />
-        </>
-    );
-};
-
-export const TitleMenu = styled(TitleMenuImpl)`
-
-`;
-
 const Container = styled.div`
     height: 100%;
     width: 100%;
@@ -192,16 +82,6 @@ const Container = styled.div`
     align-items: center;
 
     box-shadow: inset 0px -1px 0px rgba(var(--center-channel-color-rgb), 0.16);
-`;
-
-const Title = styled.div`
-    ${SemiBoldHeading}
-    letter-spacing: -0.01em;
-    font-size: 16px;
-    line-height: 24px;
-    color: var(--center-channel-color);
-    margin: 0;
-    white-space: nowrap;
 `;
 
 const Icon = styled.i`
